Frameworks: Cleanup google traslations
it is really at the moment a nightmare to see compile errors when
getting 100000 of warnings. Whatever we removed already on some
packages the warnings....but still a lot are there

This commit cleans up SystemUI

Change-Id: I4e246e1cfc13bb18eb75624510e53731821ff553

fb: google please cleanup your translations....fb cleanup

it is really at the moment a nightmare to see compile errors when
getting 100000 of warnings. Whatever we removed already on some
packages the warnings....but still a lot are there

Change-Id: I1cc3abee5b4820d9c56b89056f1ea7e4d798947b

Keyguard: google please cleanup your translations....Keyguard cleanup

it is really at the moment a nightmare to see compile errors when
getting 100000 of warnings. Whatever we removed already on some
packages the warnings....

this package is the best.......nearly every language has left
overs where not one string is used at all and one string
which is used does not exist in system language

whatever remove and add it like it should

we are now warning free

Change-Id: I06a3a3d15560aa04abd8c6c4981ac12c6dbc0c90

DocumentsUI: google please cleanup your translations....DocumentsUI cleanup

it is really at the moment a nightmare to see compile errors when
getting 100000 of warnings. Whatever we removed already on some
packages the warnings....but still a lot are there

Change-Id: I51682d5b74f76cb5e14d0a51d7694ee7235236be